PBD-3139 Move population of Derived Service Dependencies into DerivedServiceDependenciesRepository and add the slug-deny-list filtering into the Deployments lookup
